refactor(frontend): Rename SnippetForm to SnippetCreationForm
- Rename SnippetForm.tsx to SnippetCreationForm.tsx
- Update imports in page.tsx and other affected files

This change better reflects the component's purpose of creating new snippets.
