+++ frontend/src/app/components/SnippetCreationForm.tsx
